chore: delete all daemon-loader modules and remove Karaf feature definitions
Daemon-loaders were Karaf OSGi launchers from the pre-Spring-Boot era.
All daemons now run as standalone Spring Boot apps (daemon-boot-*).
Registry classes were extracted to daemon-common in the previous commit.

- Remove 7 active daemon-loader modules from core/pom.xml reactor
- Delete all 15 daemon-loader directories (7 active + 8 stubs)
- Remove daemon-loader dependencies from daemon-common, all daemon-boot
  POMs, and container/features/pom.xml
- Move InlineIdentity to daemon-boot-pollerd (same package)
- Move KafkaSinkBridge, LocalMessageDispatcherFactory, and
  TelemetryMessageConsumerManager to daemon-boot-telemetryd
- Remove all daemon-loader Karaf feature definitions from features.xml,
  both overlay features.xml files, and .boot files
