+++ core/daemon-boot-provisiond/pom.xml
@@ -403,13 +403,6 @@
             <version>${project.version}</version>
         </dependency>
 
-        <!-- Shared daemon-loader (LocalServiceDetectorRegistry) -->
-        <dependency>
-            <groupId>org.opennms.core</groupId>
-            <artifactId>org.opennms.core.daemon-loader-shared</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-
         <!-- MATE API (EntityScopeProvider) -->
         <dependency>
             <groupId>org.opennms.core.mate</groupId>
