+++ core/daemon-boot-telemetryd/src/main/java/org/opennms/netmgt/telemetry/boot/KafkaSinkBridge.java
@@ -19,7 +19,7 @@
  * language governing permissions and limitations under the
  * License.
  */
-package org.opennms.core.daemon.loader;
+package org.opennms.netmgt.telemetry.boot;
 
 import java.time.Duration;
 import java.util.Collections;
