--- core/daemon-loader-bsmd/pom.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.opennms</groupId>
-        <artifactId>org.opennms.core</artifactId>
-        <version>36.0.0-SNAPSHOT</version>
-    </parent>
-
-    <groupId>org.opennms.core</groupId>
-    <artifactId>org.opennms.core.daemon-loader-bsmd</artifactId>
-    <packaging>bundle</packaging>
-    <name>OpenNMS :: Core :: Daemon Loader :: Bsmd</name>
-    <description>
-        Karaf-only Bsmd loader. Creates and starts the Bsmd daemon in a
-        Karaf container without the Manager/Eventd, wiring all dependencies
-        directly from the OSGi service registry.
-    </description>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-RequiredExecutionEnvironment>JavaSE-17</Bundle-RequiredExecutionEnvironment>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Version>${project.version}</Bundle-Version>
-                        <Import-Package>*;resolution:=optional</Import-Package>
-                        <DynamicImport-Package>*</DynamicImport-Package>
-                        <Spring-Context>
-                            META-INF/opennms/*.xml;publish-context:=false;create-asynchronously:=true
-                        </Spring-Context>
-                    </instructions>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-    <dependencies>
-        <!-- Daemon base (SpringServiceDaemon, DaemonLifecycleManager) -->
-        <dependency>
-            <groupId>org.opennms.core</groupId>
-            <artifactId>org.opennms.core.daemon</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- BSM daemon class (Bsmd) -->
-        <dependency>
-            <groupId>org.opennms.features.bsm</groupId>
-            <artifactId>org.opennms.features.bsm.daemon</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- BSM service API (BusinessServiceStateMachine, BusinessServiceManager) -->
-        <dependency>
-            <groupId>org.opennms.features.bsm</groupId>
-            <artifactId>org.opennms.features.bsm.service.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Alarmd daemon (AlarmLifecycleListenerManager) -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>opennms-alarmd</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Alarm API (AlarmLifecycleListener) -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>opennms-alarm-api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Event API (EventIpcManager, EventSubscriptionService) -->
-        <dependency>
-            <groupId>org.opennms.features.events</groupId>
-            <artifactId>org.opennms.features.events.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- MessageBus API -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>org.opennms.core.messagebus.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- DAO APIs -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>opennms-dao-api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Spring (for InitializingBean, DisposableBean) -->
-        <dependency>
-            <groupId>org.apache.servicemix.bundles</groupId>
-            <artifactId>org.apache.servicemix.bundles.spring-beans</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Logging -->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-    </dependencies>
-</project>
