--- core/daemon-loader-bsmd/src/main/resources/META-INF/opennms/applicationContext-daemon-loader-bsmd.xml
@@ -1,133 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Karaf-only Bsmd loader context.
-
-  In a Karaf-only container (no Manager, no Eventd), this context wires Bsmd
-  directly to OSGi services.
-
-  CROSS-CONTAINER ALARM NOTIFICATION:
-  AlarmLifecycleListenerManager polls AlarmDao.findAll() every 30 seconds
-  and pushes snapshots to Bsmd (registered as local AlarmLifecycleListener).
-  This replaces same-JVM callbacks from Alarmd.
-
-  IRON RULE: Events are NEVER written to PostgreSQL. Events flow via Kafka only.
--->
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:osgi="http://www.springframework.org/schema/osgi"
-       xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-       xsi:schemaLocation="
-           http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-           http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-           http://www.springframework.org/schema/osgi http://www.springframework.org/schema/osgi/spring-osgi.xsd
-           http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-       ">
-
-    <context:annotation-config />
-
-    <osgi:reference id="serviceRegistry"
-                    interface="org.opennms.core.soa.ServiceRegistry"/>
-
-    <!-- ============================================================== -->
-    <!-- OSGi Service References                                        -->
-    <!-- ============================================================== -->
-
-    <!-- Event infrastructure (from opennms-event-forwarder-kafka blueprint) -->
-    <osgi:reference id="eventIpcManager"
-                    interface="org.opennms.netmgt.events.api.EventIpcManager"/>
-    <osgi:reference id="eventSubscriptionService"
-                    interface="org.opennms.netmgt.events.api.EventSubscriptionService"/>
-
-    <!-- BSM services (from bsm service.impl via onmsgi:service) -->
-    <onmsgi:reference id="businessServiceStateMachine"
-                      interface="org.opennms.netmgt.bsm.service.BusinessServiceStateMachine"/>
-    <onmsgi:reference id="businessServiceManager"
-                      interface="org.opennms.netmgt.bsm.service.BusinessServiceManager"/>
-
-    <!-- EventConfDao — local instance loading from DB (same pattern as Trapd/Syslogd) -->
-    <bean id="eventConfDao" class="org.opennms.netmgt.config.DefaultEventConfDao"/>
-
-    <!-- EventConfEventDao: Hibernate DAO for event_conf_event table.
-         Registered by distributed-dao-impl (applicationContext-shared.xml). -->
-    <onmsgi:reference id="eventConfEventDao"
-                      interface="org.opennms.netmgt.dao.api.EventConfEventDao"/>
-
-    <!-- Load event definitions from DB into EventConfDao on startup.
-         Without this, EventConfDao is empty and Bsmd's reduction key
-         verification in start() will fail silently.
-         Uses a class in this bundle (not MethodInvokingBean) to avoid
-         OSGi classloader issues. -->
-    <bean id="eventConfLoader" class="org.opennms.core.daemon.loader.EventConfInitializer">
-        <constructor-arg ref="eventConfEventDao"/>
-        <constructor-arg ref="eventConfDao"/>
-    </bean>
-
-    <!-- Transaction template -->
-    <onmsgi:reference id="transactionTemplate"
-                      interface="org.springframework.transaction.support.TransactionTemplate"/>
-
-    <!-- DAOs for AlarmLifecycleListenerManager (polls alarm table) -->
-    <onmsgi:reference id="alarmDao"
-                      interface="org.opennms.netmgt.dao.api.AlarmDao"/>
-    <onmsgi:reference id="sessionUtils"
-                      interface="org.opennms.netmgt.dao.api.SessionUtils"/>
-
-    <!-- ============================================================== -->
-    <!-- Alarm Lifecycle Snapshot Polling (30-second interval)          -->
-    <!-- ============================================================== -->
-    <!-- AlarmLifecycleListenerManager runs a Timer that polls
-         AlarmDao.findAll() and pushes snapshots to all registered
-         AlarmLifecycleListener instances. In standalone container,
-         this replaces same-JVM callbacks from Alarmd.
-
-         Polling interval: set via system property
-         org.opennms.alarms.snapshot.sync.ms=30000 (JAVA_OPTS in compose) -->
-
-    <bean id="alarmLifecycleListenerManager"
-          class="org.opennms.netmgt.alarmd.AlarmLifecycleListenerManager"/>
-    <!-- Register it as AlarmEntityListener so DAO change notifications flow -->
-    <onmsgi:service interface="org.opennms.netmgt.dao.api.AlarmEntityListener"
-                    ref="alarmLifecycleListenerManager">
-        <onmsgi:service-properties>
-            <entry key="registration.export" value="true"/>
-        </onmsgi:service-properties>
-    </onmsgi:service>
-    <!-- Dynamically bind/unbind AlarmLifecycleListeners (Bsmd registers below) -->
-    <onmsgi:list id="alarmLifecycleListeners"
-                 interface="org.opennms.netmgt.alarmd.api.AlarmLifecycleListener">
-        <onmsgi:listener ref="alarmLifecycleListenerManager"
-                         bind-method="onListenerRegistered"
-                         unbind-method="onListenerUnregistered"/>
-    </onmsgi:list>
-
-    <!-- ============================================================== -->
-    <!-- Bsmd Daemon (SpringServiceDaemon)                             -->
-    <!-- ============================================================== -->
-
-    <bean id="daemon" class="org.opennms.netmgt.bsm.daemon.Bsmd"
-          depends-on="eventConfLoader"/>
-
-    <!-- Event listener adapter -->
-    <bean id="daemonListener" class="org.opennms.netmgt.events.api.AnnotationBasedEventListenerAdapter">
-        <property name="annotatedListener" ref="daemon"/>
-        <property name="eventSubscriptionService" ref="eventSubscriptionService"/>
-    </bean>
-
-    <!-- Register Bsmd as AlarmLifecycleListener — the local
-         AlarmLifecycleListenerManager will call handleAlarmSnapshot()
-         every 30 seconds with all alarms from PostgreSQL -->
-    <onmsgi:service interface="org.opennms.netmgt.alarmd.api.AlarmLifecycleListener" ref="daemon">
-        <onmsgi:service-properties>
-            <entry key="registration.export" value="true"/>
-        </onmsgi:service-properties>
-    </onmsgi:service>
-
-    <!-- Daemon Lifecycle -->
-    <bean name="daemonLifecycleManager"
-          class="org.opennms.core.daemon.loader.SpringDaemonLifecycleManager">
-        <constructor-arg ref="daemon"/>
-        <constructor-arg value="Bsmd"/>
-    </bean>
-
-</beans>
