--- core/daemon-loader-collectd/src/main/resources/META-INF/opennms/applicationContext-daemon-loader-collectd.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Karaf-only Collectd loader context.
-
-  In a Karaf-only container (no Manager, no Eventd), this context replaces the
-  multi-level Spring hierarchy with a flat context that wires Collectd beans
-  directly to OSGi services.
-
-  All dependencies come from OSGi service registry:
-  - EventIpcManager from KafkaEventIpcManagerAdapter (event-forwarder-kafka feature)
-  - DAOs from distributed DAO feature (via SOA ServiceRegistry)
-  - CollectdConfigFactory created locally (opennms-config has no Spring-Context header)
-  - ServiceCollectorRegistry / LocationAwareCollectorClient created locally
-  - PersisterFactory / ThresholdingService optional (not in minimal container)
--->
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:osgi="http://www.springframework.org/schema/osgi"
-       xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-       xsi:schemaLocation="
-           http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-           http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-           http://www.springframework.org/schema/osgi http://www.springframework.org/schema/osgi/spring-osgi.xsd
-           http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-       ">
-
-    <!-- Enable @Autowired processing for Collectd -->
-    <context:annotation-config />
-
-    <!-- Shared KafkaRpcClientFactory wiring -->
-    <import resource="classpath:kafka-rpc-client-factory.xml"/>
-
-    <!-- ServiceRegistry: import the shared instance from OSGi (registered by
-         service-registry module) rather than creating a local one with
-         <onmsgi:default-registry/> which would be a different, empty instance -->
-    <osgi:reference id="serviceRegistry"
-                    interface="org.opennms.core.soa.ServiceRegistry"/>
-
-    <!-- ============================================================== -->
-    <!-- OSGi Service References (from other Karaf features)            -->
-    <!-- ============================================================== -->
-
-    <!-- Event infrastructure (from opennms-event-forwarder-kafka blueprint) -->
-    <osgi:reference id="eventIpcManager"
-                    interface="org.opennms.netmgt.events.api.EventIpcManager"/>
-
-    <!-- Collectd configuration: created locally because the opennms-config bundle
-         has no Spring-Context header, so Spring-DM never loads its XML configs.
-         Requires collectd-configuration.xml in ${opennms.home}/etc/ -->
-    <bean id="collectdConfigFactory"
-          class="org.opennms.netmgt.config.CollectdConfigFactory"/>
-
-    <!-- Collection persistence (optional — not in minimal daemon) -->
-    <onmsgi:reference id="persisterFactory"
-                      interface="org.opennms.netmgt.collection.api.PersisterFactory"/>
-
-    <!-- Thresholding (optional — not in minimal daemon) -->
-    <onmsgi:reference id="thresholdingService"
-                      interface="org.opennms.netmgt.threshd.api.ThresholdingService"/>
-
-    <!-- Entity scope provider (optional — MATE engine) -->
-    <onmsgi:reference id="entityScopeProvider"
-                      interface="org.opennms.core.mate.api.EntityScopeProvider"/>
-
-    <osgi:reference id="distPollerDao"
-                    interface="org.opennms.netmgt.dao.api.DistPollerDao"/>
-
-    <!-- DAOs (from opennms-distributed-core-impl / opennms-persistence via SOA ServiceRegistry) -->
-    <onmsgi:reference id="nodeDao"
-                      interface="org.opennms.netmgt.dao.api.NodeDao"/>
-    <onmsgi:reference id="ipInterfaceDao"
-                      interface="org.opennms.netmgt.dao.api.IpInterfaceDao"/>
-
-    <!-- Filter DAO (from opennms-config feature) -->
-    <onmsgi:reference id="filterDao"
-                      interface="org.opennms.netmgt.filter.api.FilterDao"/>
-
-    <!-- Poll outages DAO (optional — not in minimal daemon) -->
-    <onmsgi:reference id="pollOutagesDao"
-                      interface="org.opennms.netmgt.config.dao.outages.api.ReadablePollOutagesDao"/>
-
-    <!-- Resource types DAO (needed by DefaultResourceTypeMapper @Autowired) -->
-    <onmsgi:reference id="resourceTypesDao"
-                      interface="org.opennms.netmgt.config.api.ResourceTypesDao"/>
-
-    <!-- Transaction management (from opennms-persistence via SOA ServiceRegistry) -->
-    <onmsgi:reference id="platformTransactionManager"
-                      interface="org.springframework.transaction.PlatformTransactionManager"/>
-    <bean id="transactionTemplate"
-          class="org.springframework.transaction.support.TransactionTemplate">
-        <constructor-arg ref="platformTransactionManager"/>
-    </bean>
-
-    <!-- ============================================================== -->
-    <!-- Local Beans (from daemon context, normally provided by Manager) -->
-    <!-- ============================================================== -->
-
-    <!-- Service collector registry (discovers collectors via ServiceLoader) -->
-    <bean id="serviceCollectorRegistry"
-          class="org.opennms.core.daemon.loader.LocalServiceCollectorRegistry"/>
-    <onmsgi:service interface="org.opennms.netmgt.collection.api.ServiceCollectorRegistry"
-                    ref="serviceCollectorRegistry"/>
-
-    <bean id="collectorExecutor" class="java.util.concurrent.Executors"
-          factory-method="newCachedThreadPool"/>
-
-    <!-- Collector RPC module: serializes/deserializes collector request/response DTOs -->
-    <bean id="collectorClientRpcModule"
-          class="org.opennms.netmgt.collection.client.rpc.CollectorClientRpcModule"/>
-
-    <!-- Real LocationAwareCollectorClient — delegates collections to Minions via Kafka RPC -->
-    <bean id="locationAwareCollectorClient"
-          class="org.opennms.netmgt.collection.client.rpc.LocationAwareCollectorClientImpl"/>
-    <onmsgi:service interface="org.opennms.netmgt.collection.api.LocationAwareCollectorClient"
-                    ref="locationAwareCollectorClient"/>
-
-    <bean id="collectionAgentFactory"
-          class="org.opennms.netmgt.collectd.DefaultSnmpCollectionAgentFactory"/>
-    <onmsgi:service interface="org.opennms.netmgt.collection.api.CollectionAgentFactory"
-                    ref="collectionAgentFactory"/>
-
-    <bean id="defaultResourceTypeMapper"
-          class="org.opennms.netmgt.collectd.DefaultResourceTypeMapper"/>
-
-    <!-- ============================================================== -->
-    <!-- Collectd Daemon                                                -->
-    <!-- ============================================================== -->
-
-    <bean id="daemon" class="org.opennms.netmgt.collectd.Collectd">
-        <property name="eventIpcManager" ref="eventIpcManager"/>
-    </bean>
-
-    <!-- ============================================================== -->
-    <!-- Daemon Lifecycle (replaces Manager)                            -->
-    <!-- ============================================================== -->
-
-    <bean id="daemonLifecycleManager"
-          class="org.opennms.core.daemon.loader.DaemonLifecycleManager">
-        <constructor-arg ref="daemon"/>
-    </bean>
-
-</beans>
