--- core/daemon-loader-enlinkd/pom.xml
@@ -1,165 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>org.opennms</groupId>
-        <artifactId>org.opennms.core</artifactId>
-        <version>36.0.0-SNAPSHOT</version>
-    </parent>
-
-    <groupId>org.opennms.core</groupId>
-    <artifactId>org.opennms.core.daemon-loader-enlinkd</artifactId>
-    <packaging>bundle</packaging>
-    <name>OpenNMS :: Core :: Daemon Loader :: Enlinkd</name>
-    <description>
-        Karaf-only Enlinkd loader. Creates and starts the EnhancedLinkd daemon in a
-        Karaf container without the Manager/Eventd, wiring all dependencies
-        directly from the OSGi service registry.
-    </description>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-RequiredExecutionEnvironment>JavaSE-17</Bundle-RequiredExecutionEnvironment>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Version>${project.version}</Bundle-Version>
-                        <DynamicImport-Package>*</DynamicImport-Package>
-                        <Spring-Context>
-                            META-INF/opennms/*.xml;publish-context:=false;create-asynchronously:=true
-                        </Spring-Context>
-                    </instructions>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-    <dependencies>
-        <!-- Daemon base (AbstractServiceDaemon, DaemonLifecycleManager) -->
-        <dependency>
-            <groupId>org.opennms.core</groupId>
-            <artifactId>org.opennms.core.daemon</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- EnhancedLinkd daemon class -->
-        <dependency>
-            <groupId>org.opennms.features.enlinkd</groupId>
-            <artifactId>org.opennms.features.enlinkd.daemon</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Enlinkd API (ReloadableTopologyDaemon) -->
-        <dependency>
-            <groupId>org.opennms.features.enlinkd</groupId>
-            <artifactId>org.opennms.features.enlinkd.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Enlinkd config (EnhancedLinkdConfigFactory) -->
-        <dependency>
-            <groupId>org.opennms.features.enlinkd</groupId>
-            <artifactId>org.opennms.features.enlinkd.config</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Enlinkd service API (topology service interfaces) -->
-        <dependency>
-            <groupId>org.opennms.features.enlinkd</groupId>
-            <artifactId>org.opennms.features.enlinkd.service.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Event API (EventIpcManager, EventSubscriptionService) -->
-        <dependency>
-            <groupId>org.opennms.features.events</groupId>
-            <artifactId>org.opennms.features.events.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- MessageBus API -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>org.opennms.core.messagebus.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- DAO APIs -->
-        <dependency>
-            <groupId>org.opennms</groupId>
-            <artifactId>opennms-dao-api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- SNMP API (LocationAwareSnmpClient) -->
-        <dependency>
-            <groupId>org.opennms.core.snmp</groupId>
-            <artifactId>org.opennms.core.snmp.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Topologies service API (OnmsTopologyDao) -->
-        <dependency>
-            <groupId>org.opennms.features.topologies</groupId>
-            <artifactId>org.opennms.features.topologies.service.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Spring (for InitializingBean, DisposableBean) -->
-        <dependency>
-            <groupId>org.apache.servicemix.bundles</groupId>
-            <artifactId>org.apache.servicemix.bundles.spring-beans</artifactId>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Shared KafkaRpcClientFactory wiring -->
-        <dependency>
-            <groupId>org.opennms.core</groupId>
-            <artifactId>org.opennms.core.daemon-loader-shared</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <!-- SNMP RPC implementation -->
-        <dependency>
-            <groupId>org.opennms.core.snmp</groupId>
-            <artifactId>org.opennms.core.snmp.proxy.rpc-impl</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <!-- Kafka RPC -->
-        <dependency>
-            <groupId>org.opennms.core.ipc.rpc</groupId>
-            <artifactId>org.opennms.core.ipc.rpc.kafka</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.opennms.core.ipc.rpc</groupId>
-            <artifactId>org.opennms.core.ipc.rpc.api</artifactId>
-            <version>${project.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Logging -->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-        </dependency>
-    </dependencies>
-</project>
