--- core/daemon-loader-pollerd/src/main/java/org/opennms/core/daemon/loader/InlineIdentity.java
@@ -1,45 +1,0 @@
-/*
- * Licensed to The OpenNMS Group, Inc (TOG) under one or more
- * contributor license agreements.  See the LICENSE.md file
- * distributed with this work for additional information
- * regarding copyright ownership.
- *
- * TOG licenses this file to You under the GNU Affero General
- * Public License Version 3 (the "License") or (at your option)
- * any later version.  You may not use this file except in
- * compliance with the License.  You may obtain a copy of the
- * License at:
- *
- *      https://www.gnu.org/licenses/agpl-3.0.txt
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied.  See the License for the specific
- * language governing permissions and limitations under the
- * License.
- */
-package org.opennms.core.daemon.loader;
-
-import org.opennms.distributed.core.api.Identity;
-
-/**
- * Simple Identity implementation for standalone daemon containers.
- * Returns "pollerd" as the system ID and "Default" as the location.
- */
-public class InlineIdentity implements Identity {
-    @Override
-    public String getId() {
-        return "pollerd";
-    }
-
-    @Override
-    public String getLocation() {
-        return "Default";
-    }
-
-    @Override
-    public String getType() {
-        return "OpenNMS";
-    }
-}
