--- core/daemon-loader-pollerd/src/main/resources/META-INF/services/org.opennms.netmgt.poller.ServiceMonitor
@@ -1,1 +1,0 @@
-org.opennms.netmgt.poller.monitors.PassiveServiceMonitor
