+++ core/pom.xml
@@ -77,13 +77,6 @@
     <module>tracing</module>
     <module>grpc</module>
     <module>encrypt-util</module>
-    <module>daemon-loader-shared</module>
-    <module>daemon-loader-collectd</module>
-    <module>daemon-loader-pollerd</module>
-    <module>daemon-loader-enlinkd</module>
-    <module>daemon-loader-bsmd</module>
-    <module>daemon-loader-perspectivepoller</module>
-    <module>daemon-loader-telemetryd</module>
     <module>event-forwarder-kafka</module>
     <module>fileutils</module>
     <module>tsid</module>
