+++ opennms-container/delta-v/bsmd-overlay/etc/featuresBoot.d/bsmd.boot
@@ -1,2 +1,1 @@
-opennms-daemon-bsmd
 opennms-health-rest-service
