+++ opennms-container/delta-v/collectd-daemon-overlay/etc/featuresBoot.d/collectd.boot
@@ -1,2 +1,1 @@
-opennms-daemon-collectd
 opennms-health-rest-service
