+++ opennms-container/delta-v/enlinkd-overlay/etc/featuresBoot.d/enlinkd.boot
@@ -1,2 +1,1 @@
-opennms-daemon-enlinkd
 opennms-health-rest-service
