+++ opennms-container/delta-v/pollerd-daemon-overlay/etc/featuresBoot.d/pollerd.boot
@@ -1,2 +1,1 @@
-opennms-daemon-pollerd
 opennms-health-rest-service
